bot/requirements.txt
SQLAlchemy==2.0.41
aiogram==3.20.0.post0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
asyncpg==0.30.0